Refactored gryo serialization tests to have better re-use.
